+++ packages/core/src/util/index.ts
@@ -7,3 +7,4 @@ export * from './coalesceValueQuantities';
 export * from './subtractValueQuantities';
 export * from './nativeScript';
 export * as metadatum from './metadatum';
+export * from './time';
